Run on the Pi rig: real sensors, and stop shadowing stdlib secrets
- nimbus_cam/secrets.py -> keys.py: a module of that name next to the app broke
  numpy on the Pi (numpy.random imports the stdlib secrets).
- Readings are cast to plain floats; numpy scalars do not survive json.dumps.
- README: installing and running on the Pi, including libportaudio2 for the mic.

Verified on the Pi: temperature and motion from the MLX90640 over I2C, light
from the camera, sound from the webcam mic, humidity and wind from the web, and
a full Real capture rendered on the Pi itself.

+++ device/nimbus_cam/hw.py
@@ -152,7 +152,7 @@ class PiSensors:
         if f is None:
             return None
         mean = float(np.asarray(f[::8, ::8], np.float32).mean()) / 255
-        return round(10 ** (1.2 + 2.6 * mean), 0)
+        return round(float(10 ** (1.2 + 2.6 * mean)), 0)
 
     def _db(self) -> float | None:
         try:
@@ -165,7 +165,8 @@ class PiSensors:
                         break
             rec = sd.rec(4000, samplerate=16000, channels=1, dtype="float32", device=device, blocking=True)
             rms = float(np.sqrt(np.mean(np.square(rec))))
-            return round(94 + 20 * np.log10(max(rms, 1e-6)), 1)     # dBFS → rough dBA
+            # float(): numpy scalars do not survive json.dumps, and these readings are posted as JSON
+            return round(float(94 + 20 * np.log10(max(rms, 1e-6))), 1)   # dBFS → rough dBA
         except Exception:
             return None
